+++ wandb/sklearn/__init__.py
@@ -22,15 +22,15 @@ __all__ = [
     "plot_regressor",
     "plot_summary_metrics",
     "plot_learning_curve",
-    "plot_roc",
-    "plot_confusion_matrix",
-    "plot_precision_recall",
     "plot_feature_importances",
-    "plot_elbow_curve",
-    "plot_silhouette",
     "plot_class_proportion",
     "plot_calibration_curve",
-    "plot_outlier_candidates",
-    "plot_residuals",
+    "plot_roc",
+    "plot_precision_recall",
+    "plot_confusion_matrix",
     "plot_decision_boundaries",
+    "plot_elbow_curve",
+    "plot_silhouette",
+    "plot_residuals",
+    "plot_outlier_candidates",
 ]
