--- wandb/sklearn/calculate.py
@@ -1,539 +1,0 @@
-"""Calculates metrics and structures results for logging with wandb."""
-import itertools
-import time
-from warnings import simplefilter
-
-import numpy as np
-import sklearn
-from sklearn.base import clone
-from sklearn import model_selection
-from sklearn import metrics
-from sklearn.utils.multiclass import unique_labels
-
-import wandb
-from wandb.sklearn import utils
-
-from fakemodule import validate_labels  # TODO: fix
-
-# ignore all future warnings
-simplefilter(action="ignore", category=FutureWarning)
-
-CHART_LIMIT = 1000
-
-
-def summary_metrics(model=None, X=None, y=None, X_test=None, y_test=None):
-    """Calculates summary metrics for both regressors and classifiers.
-
-    Called by plot_summary_metrics to visualize metrics. Please use the function
-    plot_summary_metric() if you wish to visualize your summary metrics.
-    """
-    if (
-        utils.test_missing(model=model, X=X, y=y, X_test=X_test, y_test=y_test)
-        and utils.test_types(model=model, X=X, y=y, X_test=X_test, y_test=y_test)
-        and utils.test_fitted(model)
-    ):
-        y = np.asarray(y)
-        y_test = np.asarray(y_test)
-        metric_name = []
-        metric_value = []
-        model_name = model.__class__.__name__
-
-        params = {}
-        # Log model params to wandb.config
-        for v in vars(model):
-            if (
-                isinstance(getattr(model, v), str)
-                or isinstance(getattr(model, v), bool)
-                or isinstance(getattr(model, v), int)
-                or isinstance(getattr(model, v), float)
-            ):
-                params[v] = getattr(model, v)
-
-        # Classifier Metrics
-        if sklearn.base.is_classifier(model):
-            y_pred = model.predict(X_test)
-
-            metric_name.append("accuracy_score")
-            metric_value.append(
-                utils.round_2(sklearn.metrics.accuracy_score(y_test, y_pred))
-            )
-            metric_name.append("precision")
-            metric_value.append(
-                utils.round_2(
-                    sklearn.metrics.precision_score(y_test, y_pred, average="weighted")
-                )
-            )
-            metric_name.append("recall")
-            metric_value.append(
-                utils.round_2(
-                    sklearn.metrics.recall_score(y_test, y_pred, average="weighted")
-                )
-            )
-            metric_name.append("f1_score")
-            metric_value.append(
-                utils.round_2(
-                    sklearn.metrics.f1_score(y_test, y_pred, average="weighted")
-                )
-            )
-
-        # Regression Metrics
-        elif sklearn.base.is_regressor(model):
-            y_pred = model.predict(X_test)
-
-            metric_name.append("mae")
-            metric_value.append(
-                utils.round_2(sklearn.metrics.mean_absolute_error(y_test, y_pred))
-            )
-            metric_name.append("mse")
-            metric_value.append(
-                utils.round_2(sklearn.metrics.mean_squared_error(y_test, y_pred))
-            )
-            metric_name.append("r2_score")
-            metric_value.append(utils.round_2(sklearn.metrics.r2_score(y_test, y_pred)))
-
-        return wandb.visualize(
-            "wandb/metrics/v1",
-            wandb.Table(
-                columns=["metric_name", "metric_value", "model_name"],
-                data=[
-                    [metric_name[i], metric_value[i], model_name]
-                    for i in range(len(metric_name))
-                ],
-            ),
-        )
-
-
-def learning_curve(
-    model,
-    X,
-    y,
-    cv=None,
-    shuffle=False,
-    random_state=None,
-    train_sizes=None,
-    n_jobs=1,
-    scoring=None,
-):
-    """Trains model on datasets of varying size and generates plot of score vs size.
-
-    Called by plot_learning_curve to visualize learning curve. Please use the function
-    plot_learning_curve() if you wish to visualize your learning curves.
-    """
-    if train_sizes is None:
-        train_sizes = np.linspace(0.1, 1.0, 5)
-    if utils.test_missing(model=model, X=X, y=y) and utils.test_types(
-        model=model, X=X, y=y
-    ):
-        y = np.asarray(y)
-        train_sizes, train_scores, test_scores = model_selection.learning_curve(
-            model,
-            X,
-            y,
-            cv=cv,
-            n_jobs=n_jobs,
-            train_sizes=train_sizes,
-            scoring=scoring,
-            shuffle=shuffle,
-            random_state=random_state,
-        )
-        train_scores_mean = np.mean(train_scores, axis=1)
-        test_scores_mean = np.mean(test_scores, axis=1)
-
-        return make_learning_curve_table(
-            train_scores_mean, test_scores_mean, train_sizes
-        )
-
-
-def make_learning_curve_table(train, test, trainsize):
-    data = []
-    for i in range(len(train)):
-        if utils.check_against_limit(i, CHART_LIMIT / 2, "learning_curve"):
-            break
-        train_set = ["train", round(train[i], 2), trainsize[i]]
-        test_set = ["test", round(test[i], 2), trainsize[i]]
-        data.append(train_set)
-        data.append(test_set)
-    return wandb.visualize(
-        "wandb/learning_curve/v1",
-        wandb.Table(columns=["dataset", "score", "train_size"], data=data),
-    )
-
-
-def confusion_matrix(
-    y_true=None,
-    y_pred=None,
-    labels=None,
-    true_labels=None,
-    pred_labels=None,
-    title=None,
-    normalize=False,
-    hide_zeros=False,
-    hide_counts=False,
-):
-    """Computes the confusion matrix to evaluate the performance of a classification.
-
-    Called by plot_confusion_matrix to visualize roc curves. Please use the function
-    plot_confusion_matrix() if you wish to visualize your confusion matrix.
-    """
-    y_true = np.asarray(y_true)
-    y_pred = np.asarray(y_pred)
-
-    if utils.test_missing(y_true=y_true, y_pred=y_pred) and utils.test_types(
-        y_true=y_true, y_pred=y_pred
-    ):
-        cm = metrics.confusion_matrix(y_true, y_pred)
-        if labels is None:
-            classes = unique_labels(y_true, y_pred)
-        else:
-            classes = np.asarray(labels)
-
-        if normalize:
-            cm = cm.astype("float") / cm.sum(axis=1)[:, np.newaxis]
-            cm = np.around(cm, decimals=2)
-            cm[np.isnan(cm)] = 0.0
-
-        if true_labels is None:
-            true_classes = classes
-        else:
-            validate_labels(classes, true_labels, "true_labels")
-
-            true_label_indexes = np.in1d(classes, true_labels)
-
-            true_classes = classes[true_label_indexes]
-            cm = cm[true_label_indexes]
-
-        if pred_labels is None:
-            pred_classes = classes
-        else:
-            validate_labels(classes, pred_labels, "pred_labels")
-
-            pred_label_indexes = np.in1d(classes, pred_labels)
-
-            pred_classes = classes[pred_label_indexes]
-            cm = cm[:, pred_label_indexes]
-
-        return make_confusion_matrix_table(cm, pred_classes, true_classes, labels)
-
-
-def make_confusion_matrix_table(cm, pred_classes, true_classes, labels):
-    data, count = [], 0
-    for i, j in itertools.product(range(cm.shape[0]), range(cm.shape[1])):
-        if labels is not None and (
-            isinstance(pred_classes[i], int) or isinstance(pred_classes[0], np.integer)
-        ):
-            pred_dict = labels[pred_classes[i]]
-            true_dict = labels[true_classes[j]]
-        else:
-            pred_dict = pred_classes[i]
-            true_dict = true_classes[j]
-        data.append([pred_dict, true_dict, cm[i, j]])
-        count += 1
-        if utils.check_against_limit(count, CHART_LIMIT, "confusion_matrix"):
-            break
-    return wandb.visualize(
-        "wandb/confusion_matrix/v1",
-        wandb.Table(columns=["Predicted", "Actual", "Count"], data=data),
-    )
-
-
-# Draw a stem plot with the influence for each instance
-# format:
-# x = feature_names[:max_num_features]
-# y = importances[indices][:max_num_features]
-def make_feature_importances_table(feature_names, importances):
-    return wandb.visualize(
-        "wandb/feature_importances/v1",
-        wandb.Table(
-            columns=["feature_names", "importances"],
-            data=[],
-        ),
-    )
-
-
-def get_attributes_as_formatted_string(attributes):
-    result = ""
-    for index in range(len(attributes) - 1):
-        if result == "":
-            result = attributes[index]
-        else:
-            result = ", ".join([result, attributes[index]])
-
-    return " or ".join([result, attributes[-1]])
-
-
-def _clone_and_score_clusterer(clusterer, X, n_clusters):
-    start = time.time()
-    clusterer = clone(clusterer)
-    setattr(clusterer, "n_clusters", n_clusters)
-    return clusterer.fit(X).score(X), time.time() - start
-
-
-def make_elbow_curve_table(cluster_ranges, clfs, times):
-    return wandb.visualize(
-        "wandb/elbow/v1",
-        wandb.Table(
-            columns=["cluster_ranges", "errors", "clustering_time"],
-            data=[
-                [cluster_ranges[i], clfs[i], times[i]]
-                for i in range(len(cluster_ranges))
-            ],
-        ),
-    )
-
-
-def silhouette(x, y, colors, centerx, centery, y_sil, x_sil, color_sil, silhouette_avg):
-    return wandb.visualize(
-        "wandb/silhouette_/v1",
-        wandb.Table(
-            columns=[
-                "x",
-                "y",
-                "colors",
-                "centerx",
-                "centery",
-                "y_sil",
-                "x1",
-                "x2",
-                "color_sil",
-                "silhouette_avg",
-            ],
-            data=[
-                [
-                    x[i],
-                    y[i],
-                    colors[i],
-                    centerx[colors[i]],
-                    centery[colors[i]],
-                    y_sil[i],
-                    0,
-                    x_sil[i],
-                    color_sil[i],
-                    silhouette_avg,
-                ]
-                for i in range(len(color_sil))
-            ],
-        ),
-    )
-
-
-def silhouette_(
-    x, y, colors, centerx, centery, y_sil, x_sil, color_sil, silhouette_avg
-):
-    return wandb.visualize(
-        "wandb/silhouette_/v1",
-        wandb.Table(
-            columns=[
-                "x",
-                "y",
-                "colors",
-                "centerx",
-                "centery",
-                "y_sil",
-                "x1",
-                "x2",
-                "color_sil",
-                "silhouette_avg",
-            ],
-            data=[
-                [
-                    x[i],
-                    y[i],
-                    colors[i],
-                    None,
-                    None,
-                    y_sil[i],
-                    0,
-                    x_sil[i],
-                    color_sil[i],
-                    silhouette_avg,
-                ]
-                for i in range(len(color_sil))
-            ],
-        ),
-    )
-
-
-def class_proportions(classes_, class_counts_train, class_counts_test, labels):
-    class_dict = []
-    dataset_dict = []
-    count_dict = []
-    for i in range(len(classes_)):
-        # add class counts from training set
-        class_dict.append(classes_[i])
-        dataset_dict.append("train")
-        count_dict.append(class_counts_train[i])
-        # add class counts from test set
-        class_dict.append(classes_[i])
-        dataset_dict.append("test")
-        count_dict.append(class_counts_test[i])
-        if utils.check_against_limit(i, CHART_LIMIT, "class_proportions"):
-            break
-
-    if labels is not None and (
-        isinstance(class_dict[0], int) or isinstance(class_dict[0], np.integer)
-    ):
-        class_dict = get_named_labels(labels, class_dict)
-    return wandb.visualize(
-        "wandb/class_proportions/v1",
-        wandb.Table(
-            columns=["class", "dataset", "count"],
-            data=[
-                [class_dict[i], dataset_dict[i], count_dict[i]]
-                for i in range(len(class_dict))
-            ],
-        ),
-    )
-
-
-def calibration_curves(
-    model_dict,
-    frac_positives_dict,
-    mean_pred_value_dict,
-    hist_dict,
-    edge_dict,
-):
-    return wandb.visualize(
-        "wandb/calibration/v1",
-        wandb.Table(
-            columns=[
-                "model",
-                "fraction_of_positives",
-                "mean_predicted_value",
-                "hist_dict",
-                "edge_dict",
-            ],
-            data=[
-                [
-                    model_dict[i],
-                    frac_positives_dict[i],
-                    mean_pred_value_dict[i],
-                    hist_dict[i],
-                    edge_dict[i],
-                ]
-                for i in range(len(model_dict))
-            ],
-        ),
-    )
-
-
-# Draw a stem plot with the influence for each instance
-# format: distance_, len(distance_), influence_threshold_, utils.round_3(outlier_percentage_)
-def outlier_candidates(distance, outlier_percentage, influence_threshold):
-    return wandb.visualize(
-        "wandb/outliers/v1",
-        wandb.Table(
-            columns=[
-                "distance",
-                "instance_indicies",
-                "outlier_percentage",
-                "influence_threshold",
-            ],
-            data=[
-                [
-                    distance[i],
-                    i,
-                    utils.round_3(outlier_percentage),
-                    influence_threshold,
-                ]
-                for i in range(len(distance))
-            ],
-        ),
-    )
-
-
-# format:
-# Legend: train_score_, test_score_ (play with opacity)
-# Scatterplot: dataset(train, test)(color), y_pred(x), residuals(y)
-# Histogram: dataset(train, test)(color), residuals(y), aggregate(residuals(x)) with bins=50
-def residuals(
-    y_pred_train,
-    residuals_train,
-    y_pred_test,
-    residuals_test,
-    train_score_,
-    test_score_,
-):
-    y_pred_dict = []
-    dataset_dict = []
-    residuals_dict = []
-    datapoints = 0
-    max_datapoints_train = 900
-    max_datapoints_train = 100
-    for pred, residual in zip(y_pred_train, residuals_train):
-        # add class counts from training set
-        y_pred_dict.append(pred)
-        dataset_dict.append("train")
-        residuals_dict.append(residual)
-        datapoints += 1
-        if utils.check_against_limit(datapoints, max_datapoints_train, "residuals"):
-            break
-    datapoints = 0
-    for pred, residual in zip(y_pred_test, residuals_test):
-        # add class counts from training set
-        y_pred_dict.append(pred)
-        dataset_dict.append("test")
-        residuals_dict.append(residual)
-        datapoints += 1
-        if utils.check_against_limit(datapoints, max_datapoints_train, "residuals"):
-            break
-
-    return wandb.visualize(
-        "wandb/residuals_plot/v1",
-        wandb.Table(
-            columns=[
-                "dataset",
-                "y_pred",
-                "residuals",
-                "train_score",
-                "test_score",
-            ],
-            data=[
-                [
-                    dataset_dict[i],
-                    y_pred_dict[i],
-                    residuals_dict[i],
-                    train_score_,
-                    test_score_,
-                ]
-                for i in range(len(y_pred_dict))
-            ],
-        ),
-    )
-
-
-def decision_boundaries(
-    decision_boundary_x,
-    decision_boundary_y,
-    decision_boundary_color,
-    train_x,
-    train_y,
-    train_color,
-    test_x,
-    test_y,
-    test_color,
-):
-    x_dict, y_dict, color_dict = [], [], []
-    for i in range(min(len(decision_boundary_x), 100)):
-        x_dict.append(decision_boundary_x[i])
-        y_dict.append(decision_boundary_y[i])
-        color_dict.append(decision_boundary_color)
-    for i in range(300):
-        x_dict.append(test_x[i])
-        y_dict.append(test_y[i])
-        color_dict.append(test_color[i])
-    for i in range(min(len(train_x), 600)):
-        x_dict.append(train_x[i])
-        y_dict.append(train_y[i])
-        color_dict.append(train_color[i])
-
-    return wandb.visualize(
-        "wandb/decision_boundaries/v1",
-        wandb.Table(
-            columns=["x", "y", "color"],
-            data=[[x_dict[i], y_dict[i], color_dict[i]] for i in range(len(x_dict))],
-        ),
-    )
-
-
-def get_named_labels(labels, numeric_labels):
-    return np.array([labels[num_label] for num_label in numeric_labels])
